Increased navigation URL size
There was an error when a user tried to add a navigation to an external website like their youtube channel. The URL would be cut off prematurely. The cause was that the database column was set to 32 characters which is not nearly enough for some URLS. The number has been increased to 255.