Fix permission Issues & Update php
This dockerfile is meant to be paired with an nginx server, which sometimes created write problems. By giving ownership to the nginx user we shouldn't have any problems.

Also updated to a smaller alpine image, and hardcoded php 7.2.

Removed the php debugger extensions as this image can be used in production.